+++ kokoro/macos/package.sh
@@ -16,7 +16,7 @@
 # MacOS Package Script.
 set -ex
 
-if [ $# -ne 1 -o ! -f "$1/pkg/build.properties" ]; then
+if [ $# -ne 1 -o ! -d "$1" ]; then
 	echo Expected the build folder as an argument.
 	exit 1
 fi
@@ -27,6 +27,12 @@ function absname {
 
 BUILD_OUT=$1
 SRC=$(absname "$(dirname "${BASH_SOURCE[0]}")")
+BIN=$SRC/../../bazel-bin
+
+if [ ! -f "$BIN/pkg/build.properties" ]; then
+  echo Unable to find pkg/build.properties in $BIN
+  exit 1
+fi
 
 rm -rf "$BUILD_OUT/dist"
 mkdir -p "$BUILD_OUT/dist"
@@ -35,12 +41,11 @@ VERSION=$(awk -F= 'BEGIN {major=0; minor=0; micro=0}
                   /Major/ {major=$2}
                   /Minor/ {minor=$2}
                   /Micro/ {micro=$2}
-                  END {print major"."minor"."micro}' ../pkg/build.properties)
+                  END {print major"."minor"."micro}' $BIN/pkg/build.properties)
 
 # Combine package contents.
 mkdir -p gapid/jre
-cp -r ../pkg/* gapid/
-cp -r ../current/java/gapic-osx.jar gapid/lib/gapic.jar
+cp -r $BIN/pkg/* gapid/
 "$SRC/copy_jre.sh" gapid/jre
 
 # Create a zip file.
@@ -58,8 +63,8 @@ for i in 512 256 128 64 32 16; do
 done
 iconutil -c icns -o GAPID.app/Contents/Resources/GAPID.icns GAPID.iconset
 
-# Copy the GAPIR Symbols
-cp ../current/gapir.sym gapir-$VERSION-macos.sym
+# TODO Copy the GAPIR Symbols
+# cp ../current/gapir.sym gapir-$VERSION-macos.sym
 
 # Make a dmg file.
 pip install --upgrade --user dmgbuild pyobjc-framework-Quartz
